setup.py: use 'version' so zest.releaser can update it.
zest.releaser cannot update the '__version__' with underscores.  We
could keep this variable, but there is no way to actually access this
variable.  If we want underscores and want to access this, it should
be in zope/globalrequest/__init__.py, something like this:

import pkg_resources
__version__ = pkg_resources.get_distribution("zope.globalrequest").version

See http://zestreleaser.readthedocs.org/en/latest/versions.html